pyproject.toml
[project]
name = "bootcamp"
version = "0.1.0"
description = ""
authors = [
    {name = "hojiakbar2806",email = "hojiakbarnasriddinov2006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram[18n] (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "babel (>=2.17.0,<3.0.0)",
    "redis-dict (>=3.2.5,<4.0.0)",
    "apscheduler (>=3.11.1,<4.0.0)",
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
